requirements.txt
python-chess
numpy
torch
python-dotenv
tqdm